Sfang/epoch boundary (#86)
* distribute reward logic with sysvar

* new bank update for epoch

* new bank update for epoch

* size up

* resolve segfault

* downsizing to pass tests

* more initialization

* clean up

* small bug fix

* adjust comments

* more structure

* remove header

* wip fix for fd_runtime_init_bank_from_genesis

* need to fix stake_history

* fix block height

* stake history sysvar fix

* remove unneeded header

* get rid of fd_stakes_init

* resolve seg faults + lamports getting closer

* resolve clang issue

* more fixes + logging

* fix clang errors

* capitalization sysvar fix

* changes for capitalization

* burn more fee

+++ src/flamenco/rewards/fd_rewards.c
@@ -252,7 +252,7 @@ epoch_duration_in_years(
     FD_LOG_NOTICE(("epoch_duration_in_years bank->epoch_schedule.first_normal_epoch: %lu", bank->epoch_schedule.first_normal_epoch));
     FD_LOG_NOTICE(("epoch_duration_in_years slots_in_epoch: %lu", slots_in_epoch));
     FD_LOG_NOTICE(("epoch_duration_in_years bank->slots_per_year: %f", bank->slots_per_year));
-    return (double)slots_in_epoch / bank->slots_per_year;
+    return (double)slots_in_epoch / (double) bank->slots_per_year;
 }
 
 static void
